Struct spellck::visitor::SpellingVisitor

pub struct SpellingVisitor<'a> {
    pub misspellings: BTreeMap<Position, Vec<String>>,
    // some fields omitted
}

Keeps track of the reference dictionary and the misspelled words through a traversal of the whole ast.

Fields

misspellings

The misspelled words

Methods

impl<'a> SpellingVisitor<'a>

fn new<'b>(words: &'b HashSet<String>, exported: &'b ExportedItems) -> SpellingVisitor<'b>

ast::Create a new Spelling Visitor.

fn check_crate(&mut self, krate: &Crate)

Spell-check a whole krate.